Restructuring the directory structure
