--- examples/train/thunder_agent/fully_async_run_gsm8k.sh
@@ -1,65 +1,0 @@
-set -x
-
-# Fully async GRPO training+generation for Qwen2.5-1.5B-Instruct on GSM8K
-# using ThunderAgent as the HTTP inference router.
-
-# uv run examples/train/gsm8k/gsm8k_dataset.py --output_dir $HOME/data/gsm8k
-# export WANDB_API_KEY=<your_key_here>
-# bash examples/train/thunder_agent/fully_async_run_gsm8k.sh
-
-: "${DATA_DIR:="$HOME/data/gsm8k"}"
-: "${NUM_INFERENCE_GPUS:=2}"
-: "${NUM_POLICY_GPUS:=2}"
-: "${LOGGER:=wandb}"
-: "${INFERENCE_BACKEND:=vllm}"
-
-: "${MINI_BATCH_SIZE:=256}"
-: "${MAX_STALENESS_STEPS:=4}"
-: "${NUM_PARALLEL_GENERATION_WORKERS:=$(( MINI_BATCH_SIZE * (MAX_STALENESS_STEPS + 1) ))}"
-
-RUN_NAME=gsm8k-fully-async-thunderagent-qwen2.5_1.5B-maxStale${MAX_STALENESS_STEPS}-${NUM_POLICY_GPUS}train${NUM_INFERENCE_GPUS}gen
-
-_SKYRL_USE_NEW_INFERENCE=1 uv run --isolated --extra fsdp -m examples.train.thunder_agent.main_fully_async_thunder_agent \
-  data.train_data="['$DATA_DIR/train.parquet']" \
-  data.val_data="['$DATA_DIR/validation.parquet']" \
-  trainer.fully_async.max_staleness_steps=${MAX_STALENESS_STEPS} \
-  trainer.fully_async.num_parallel_generation_workers=${NUM_PARALLEL_GENERATION_WORKERS} \
-  trainer.algorithm.advantage_estimator="grpo" \
-  trainer.policy.model.path="Qwen/Qwen2.5-1.5B-Instruct" \
-  trainer.placement.colocate_all=false \
-  trainer.strategy=fsdp2 \
-  trainer.placement.policy_num_gpus_per_node=$NUM_POLICY_GPUS \
-  trainer.placement.critic_num_gpus_per_node=$NUM_POLICY_GPUS \
-  trainer.placement.ref_num_gpus_per_node=$NUM_POLICY_GPUS \
-  generator.inference_engine.num_engines=$NUM_INFERENCE_GPUS \
-  generator.inference_engine.tensor_parallel_size=1 \
-  trainer.epochs=20 \
-  trainer.eval_batch_size=1024 \
-  trainer.eval_before_train=false \
-  trainer.eval_interval=4 \
-  trainer.update_epochs_per_batch=1 \
-  trainer.train_batch_size=${MINI_BATCH_SIZE} \
-  trainer.policy_mini_batch_size=${MINI_BATCH_SIZE} \
-  trainer.micro_forward_batch_size_per_gpu=8 \
-  trainer.micro_train_batch_size_per_gpu=8 \
-  trainer.ckpt_interval=10 \
-  trainer.max_prompt_length=512 \
-  generator.sampling_params.max_generate_length=1024 \
-  trainer.policy.optimizer_config.lr=1.0e-6 \
-  trainer.algorithm.use_kl_loss=true \
-  generator.inference_engine.backend=$INFERENCE_BACKEND \
-  generator.inference_engine.run_engines_locally=true \
-  generator.inference_engine.weight_sync_backend=nccl \
-  generator.inference_engine.async_engine=true \
-  generator.inference_engine.thunder_agent_mode=tr \
-  generator.batched=false \
-  environment.env_class=gsm8k \
-  generator.n_samples_per_prompt=5 \
-  generator.inference_engine.gpu_memory_utilization=0.8 \
-  trainer.logger="$LOGGER" \
-  trainer.project_name="gsm8k-thunderagent" \
-  trainer.run_name=${RUN_NAME} \
-  trainer.resume_mode=latest \
-  trainer.ckpt_path="$HOME/ckpts/${RUN_NAME}" \
-  generator.inference_engine.enforce_eager=true \
-  $@
