--- examples/train/thunder_agent/main_fully_async_thunder_agent.py
@@ -1,64 +1,0 @@
-"""Fully async ThunderAgent training entrypoint."""
-
-import asyncio
-import faulthandler
-import signal
-import sys
-
-import ray
-
-from skyrl.train.fully_async_trainer import FullyAsyncRayPPOTrainer
-from skyrl.train.utils import initialize_ray, validate_cfg
-
-from .config import ThunderAgentConfig
-from .main_thunder_agent import ThunderAgentExp
-
-
-class FullyAsyncThunderAgentExp(ThunderAgentExp):
-    """ThunderAgent entrypoint that swaps in the fully async trainer."""
-
-    def get_trainer(
-        self,
-        cfg,
-        tracker,
-        tokenizer,
-        train_dataset,
-        eval_dataset,
-        inference_engine_client,
-        generator,
-        colocate_pg,
-    ):
-        return FullyAsyncRayPPOTrainer(
-            cfg=cfg,
-            tracker=tracker,
-            tokenizer=tokenizer,
-            train_dataset=train_dataset,
-            eval_dataset=eval_dataset,
-            inference_engine_client=inference_engine_client,
-            generator=generator,
-            colocate_pg=colocate_pg,
-        )
-
-    def run(self):
-        trainer = self._setup_trainer()
-        asyncio.run(trainer.train())
-
-
-@ray.remote(num_cpus=1)
-def skyrl_entrypoint(cfg):
-    faulthandler.enable(all_threads=True)
-    if hasattr(signal, "SIGUSR1"):
-        faulthandler.register(signal.SIGUSR1, all_threads=True, chain=False)
-    exp = FullyAsyncThunderAgentExp(cfg)
-    exp.run()
-
-
-def main() -> None:
-    cfg = ThunderAgentConfig.from_cli_overrides(sys.argv[1:])
-    validate_cfg(cfg)
-    initialize_ray(cfg)
-    ray.get(skyrl_entrypoint.remote(cfg))
-
-
-if __name__ == "__main__":
-    main()
